[ui] DesktopWidget changes the contentRect during dock and undock
http://redmine.plexyplanet.org/issues/218

The content rect should not do altered internally and it should only
change when the content or the parent widget wishes to change it's size,
but when we dock and undock (double click event). we explicitly change
the content size which results in a widget never restoring to it's
original size.

closes #218