[core] Fix Multi-screen setup
http://redmine.plexyplanet.org/issues/223

Since the current design doesn't allow us to send the background size
to the item and since the current background plugin just returns a
QGraphicsPixmapItem there is no direct way to set the contentRect
of a BackgroundItem. So we change the ControllerInterface::defaultView
to always return a AbstractDesktopView insted of a QGraphicsItem. so
that we can set the rect size from the AbstractDesktopView when creating
background controllers.

refs #223