[core] Update the API to use ref counted pointers and lazy load plugins
http://redmine.plexyplanet.org/issues/243

This commit does serveral things, mainly we define different types of
interfaces for the types of plugins we have in plexydesk, this makes
plugin loading more safer so that we don't try to load plugins that are
of unknow type.Since the current single intefcace can load any QObject
without checking. This commits also makes sure we deal with shared
pointers and not pointers directly this makes memeory management much
easier. Rest of the commits updates the Plugsin to match the api change.

closes #243